+++ src/controls/Controls.jsx
@@ -4,12 +4,12 @@ import { OrbitControls } from '@react-three/drei';
 const Controls = () => {
   return (
     <OrbitControls
-      minAzimuthAngle={-Math.PI / 2} 
-      maxAzimuthAngle={Math.PI / 2}  
+      minAzimuthAngle={-Math.PI / 4} 
+      maxAzimuthAngle={Math.PI / 4}  
 
      
       minPolarAngle={0}        
-      maxPolarAngle={Math.PI / 2} 
+      maxPolarAngle={Math.PI / 2.5} 
 
       enableZoom={true} 
     />
