+++ src/pages/sitemap/Sitemap.css
@@ -1,52 +1,71 @@
+
 .sitemap-container {
-    padding: 20px;
-    font-family: Arial, sans-serif;
-    max-width: 800px; 
-    margin: 0 auto; 
-    background-color: #a0d8ec; 
-    border-radius: 8px; 
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1); 
-  }
-  
-  h1 {
-    color: #2c3e50; 
-    font-size: 2.5rem; 
-    margin-bottom: 20px; 
-    text-align: center; 
-  }
-  
-  h2 {
-    color: #34495e; 
-    font-size: 1.8rem; 
-    margin-top: 30px; 
-    margin-bottom: 10px; 
-  }
-  
-  ul {
-    list-style-type: none; 
-    padding: 0; 
-  }
-  
-  li {
-    margin: 10px 0; 
-  }
-  
-  a {
-    color: #3498db; 
-    text-decoration: none; 
-    font-size: 1.1rem; 
-    transition: color 0.3s; 
-  }
-  
-  a:hover {
-    text-decoration: underline; 
-    color: #2980b9; 
-  }
-  
-  footer {
-    margin-top: 20px; 
-    font-size: 0.9rem; 
-    color: #7f8c8d; 
-    text-align: center; 
-  }
-  
+  position: relative;
+  padding: 20px;
+  max-width: 800px;
+  margin: 0 auto;
+  background-color: #a0d8ec;
+  border-radius: 8px;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+}
+
+h1 {
+  color: #2c3e50;
+  font-size: 2.5rem;
+  margin-bottom: 20px;
+  text-align: center;
+}
+
+h2 {
+  color: #34495e;
+  font-size: 1.8rem;
+  margin-top: 30px;
+  margin-bottom: 10px;
+}
+
+ul {
+  list-style-type: none;
+  padding: 0;
+}
+
+li {
+  margin: 10px 0;
+}
+
+a {
+  color: #3498db;
+  text-decoration: none;
+  font-size: 1.1rem;
+  transition: color 0.3s;
+}
+
+a:hover {
+  text-decoration: underline;
+  color: #2980b9;
+}
+
+footer {
+  margin-top: 20px;
+  font-size: 0.9rem;
+  color: #7f8c8d;
+  text-align: center;
+}
+
+.staging-background {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 0;
+  overflow: hidden;
+}
+
+.staging-background {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100vh; 
+  z-index: 1; 
+}
